добавил логотип, вынес навигацию в отдельный компонент, поправил стили для навигации
