+++ src/components/footer/footer.module.css
@@ -18,24 +18,8 @@
     flex-direction: column;
 }
 
-.link {
-    font-size: var(--text-size);
-    text-decoration: none;
-    color: var(--text-color);
-}
-
-.active {
-    font-size: var(--text-size);
-    text-decoration: underline;
-    color: var(--text-color);
-}
-
-.link:hover {
-    text-decoration: underline;
-}
-
 .logo {
-    margin: 20px 0;
+    margin: 0;
     display: flex;
     width: 100%;
     font-family: "Inter", sans-serif;
@@ -61,28 +45,6 @@
     font-size: 18px;
     line-height: 25px;
     text-align: start;
-}
-
-.nav {
-    position: absolute;
-    display: flex;
-    min-width: var(--screen-width);
-    flex-direction: row;
-    justify-content: center;
-}
-
-.list {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    list-style: none;
-    font-family: "Open Sans", sans-serif;
-    gap: 44px;
-}
-
-.item {
-    line-height: 25px;
 }
 
 .social {
